+++ packages/core/src/connectivity/scp-cf/destination/destination-from-service.ts
@@ -114,6 +114,12 @@ class DestinationFromServiceRetriever {
     if (destination.authentication === 'OAuth2ClientCredentials') {
       destination = await da.addOAuth2ClientCredentials();
     }
+    if (destination.authentication === 'OAuth2JWTBearer') {
+      destination = await da.addOAuthSamlAuth(
+        destination,
+        destinationResult.origin
+      );
+    }
     if (destination.authentication === 'ClientCertificateAuthentication') {
       destination = await da.addClientCertAuth();
     }